release v0.9.0 - added support for 'typedef'

+++ src/headers/stdlib.h
@@ -1,8 +1,7 @@
 #ifndef _VBRCC_STDLIB_H
 #define _VBRCC_STDLIB_H
 
-/* malloc takes long, not size_t, until typedef arrives. */
-void *malloc(long size);
+void *malloc(size_t size);
 void free(void *p);
 int abs(int n);
 void exit(int code);
